Move a bunch of the dynamic data rewriting into core
This commit moves a bunch of the code that was intended to do dynamic
data rewriting/remapping into the core function, and has this code
triggered by observing certain configurations in the collection_tables.
Makes the config code more readable/manageable at the expense of extra
logic in the CRITSDataView code. In particular, we move the bulk of the
fieldToValue(x) function in there, and then allow the author to provide
small functions on a field-by-field basis for rewriting the data.